Rework undo functionality.
Rather than a decorator on a method, you now create an instance of a subclass of `UndoableCommand`, immediately passing it to `UndoController.execute()`.

This lets the command store the state required to undo & redo operations - no need for the original method to return the data required to "undo" something, and no need to clutter the class with an "undo" method which might be otherwise unnecessary.

I've removed the weakref stuff, hopefully this doesn't affect GC.

I've also removed g_undo_controller. The undo controller now lives in DTApp.

DROData.delete_multiple() no longer accepts `is_sorted`.